Add ControlPoint
The primary way to send messages to a modem is through what is called a
control point. A control point provides data on the device, service, and
the client id that is necessary for communication with the system
device.

Control points should be managed by consumers as there are times you
want to release the control point and there are times you want to hang
on to the control point. This should be an consumer level choice.

This PR introduces the control point data structure and updates the high
level `QMI` API to use it. However, I did not refactor the level low
APIs. For now we can abstract around the low level APIs until we have a
chance to refactor any of those if we want to. Doing it this way allows
me to continue supporting QMI VintageNet without having to refactor
everything. I will add and update other pieces as use cases show
themselves while implementing support in VintageNet for QMI.