Separate out serializing and request sending
This changes the service module functions into data structures that
implement a way to be encoded into binary or parse the TLV values.

The primary reason for the change is to allow more testability into the
library as it is currently hard to test since the encoding and decoding
of the commands are mixed in with the actual sending of requests.

There are a few ways of getting the separation between sending and
serializing with various trade offs. From my experience in Grizzly I
wish I would have done protocols over behaviours, however, I can see the
argument for behaviours.

If we go this route there are some more areas I can improve upon in the
API, but without these changes those updates would not be worth the
time.