Retry to open device after the dev bridge closes
We were retrying to open the port process to the dev bridge when we were
notified that there was some error, often either `:enoent` or `:enodev`.
However, the way the dev bridge works is it notifies us about the error
then it closes. There was a condition that we when tried to close the
port in after being notified of the error, the port would close before
we called `Port.close/1` causing a bad argument error.

This change waits for the closed message before trying to reopen the
port to ensure that the state of the dev bridge does not have the old
port that has closed.