Create top-level supervision tree to isolate client IDs
Client IDs are pretty important and if you run out, then nothing works.
Since it's possible for the driver to crash, we want to be able to
restart it without losing client IDs. This puts the driver and client ID
in different branches of the supervision tree so unless things get
really bad, they won't be lost.