Fix missing transaction for response
In some cases, the driver will timeout a transaction before the response
is sent from the device. When the timeout takes place, the driver
removes the transaction from the transactions list. This causes a
problem when we try to look up the transaction and it is no longer in
the list when the response is received.

This fix will log that we received a transaction response without a
corresponding transaction request.