Fix TX/RX stats when stat value is missing
When the stat value for a TX/RX stat is missing QMI reports back
`0xFFFFFFFF`, so we mark that is `0`.

This happens because the wireless data service event report was
configured to report those stats, but those stats have not be recorded
yet.

We use libqmi as a reference point for this:
https://gitlab.freedesktop.org/mobile-broadband/libqmi/-/blob/master/src/qmicli/qmicli-wds.c#L1068-1091